+++ src/bvals/comms/bnd_info.hpp
@@ -82,6 +82,8 @@ struct BndInfo {
   BndInfo(const BndInfo &) = default;
   BndInfo(MeshBlock *pmb, const NeighborBlock &nb, std::shared_ptr<Variable<Real>> v,
           CommBuffer<buf_pool_t<Real>::owner_t> *combuf, IndexRangeType idx_range_type);
+  KOKKOS_DEFAULTED_FUNCTION
+  ~BndInfo() = default;
 
   // These are are used to generate the BndInfo struct for various
   // kinds of boundary types and operations.